合并到20240319 NoStackMapTableEver！！！ A=增加, B=修复, C=更改, D=删除
A:
	一个新的ASM工具，不进行栈帧验证的同时使用lambda表达式
		TransformUtil#noStackFrameTableEver
		使用lambda代理，完美的绕过StackMapTable

	加入新的文件比较GUI
	BsDiff支持对输入的一部分生成Diff
	AE-Cross支持保存GUI配置到yml
	重构小说整理工具，并加入了转换Epub功能
	XML支持解析注释节点
	重构Zip读写相关代码，抽取一个只读的ZipFile
	加入新的Cpk压缩工具来高效的压缩jar

B:
	优化DirectAccessor#delegate模糊匹配方法的逻辑
	Latin1字符范围是0-255而不是127，删除之前的fast path
	改进Vinary格式的读取性能
	修复MSS协议的bug
	LinkedMyHashMap哈希分布太密集时会栈溢出
	LRUCache无法正常移除缓存
	部分YAML无法解析
	MyDataInputStream的readFully和readAscii不能使用
	HTTP:
		池化的请求被远程终止后可能导致请求失败
		closeGracefully可能会在标准错误上打印异常
	MyPassIs的参数验证不够全面
	在包装的AnsiOut.print函数调用后，被覆盖的BottomLine无法正确的刷新
	循环定时任务无法取消
	asm
		修复了CstTop计算哈希值的bug
		ConstantData#getBytes支持非空的DynByteBuf
		几个toString函数可能产生异常或丢失信息的问题
		修复Nixim之前改包名忘改的部分

C:
	优化getMin2PowerOf函数
	优化7z多线程解压的性能
		source不再缓冲，因为应该没人用int read()
	优化DynByteBuf字符串处理性能，并为IDEA补全优化
	用时间轮重构Scheduler
	重构KeyFile
	铲掉一坨NovelFrame中的**
	AE-Cross使用更多的logger替换println
	MyChat3，正在用Vue写前端
	优化roj.asmx.launcher的资源加载性能，不再使用URLClassLoader
	重构CRC32，支持硬件加速
	重构TextUtil#split和XMLParser，它们可以按照我的设计处理字符串结尾的空格了

D:
	删除FMD大部分功能
	删除很多不再使用的类

+++ java/roj/io/MyRegionFile.java
@@ -99,7 +99,7 @@ public class MyRegionFile implements AutoCloseable {
 			int off = n >>> 8;
 			int len = n & 255;
 			if (len == 255 && off <= sectors) {
-				raf.seek(off * chunkSize);
+				raf.seek((long) off * chunkSize);
 				len = chunkCount(raf.asDataInput().readInt() + 4);
 			}
 
@@ -142,12 +142,12 @@ public class MyRegionFile implements AutoCloseable {
 
 		assert off + len < sectorCount;
 
-		raf.seek(off*chunkSize);
+		raf.seek((long) off * chunkSize);
 		int byteLength = raf.asDataInput().readInt();
 		if (byteLength > chunkSize*len - 4 && len != 255) {
 			log("无效的块: {} #{} 范围:[{}+{}] 数据溢出(Block): {} > {}", file, id, off, len, byteLength, len*chunkSize - 4);
 			return null;
-		} else if (byteLength + off*chunkSize > raf.length()) {
+		} else if (byteLength + (long) off * chunkSize > raf.length()) {
 			log("无效的块: {} #{} 范围:[{}+{}] 数据溢出(Global): {} > {}", file, id, off, len, byteLength + off*chunkSize, raf.length());
 			return null;
 		} else {
@@ -181,7 +181,7 @@ public class MyRegionFile implements AutoCloseable {
 		int off = i1 >>> 8;
 		int oldCLen = i1 & 255;
 		if (oldCLen == 255) {
-			raf.seek(off*chunkSize);
+			raf.seek((long) off * chunkSize);
 			oldCLen = chunkCount(raf.asDataInput().readInt() + 4);
 		}
 
@@ -232,7 +232,7 @@ public class MyRegionFile implements AutoCloseable {
 			// 在文件后增加
 			int extraBlocks = cLen + lastAvailBlock;
 
-			raf.setLength(chunkSize*extraBlocks);
+			raf.setLength((long) chunkSize * extraBlocks);
 
 			int sec = sectorCount;
 			sectorCount = extraBlocks;
@@ -259,7 +259,7 @@ public class MyRegionFile implements AutoCloseable {
 		int len = i & 255;
 
 		if (len == 255) {
-			raf.seek(off * chunkSize);
+			raf.seek((long) off * chunkSize);
 			len = chunkCount(raf.asDataInput().readInt() + 4);
 		}
 
@@ -350,7 +350,7 @@ public class MyRegionFile implements AutoCloseable {
 		writeBlock(raf, off, data);
 	}
 	private void writeBlock(Source raf, int off, DynByteBuf data) throws IOException {
-		raf.seek(off*chunkSize);
+		raf.seek((long) off * chunkSize);
 		raf.writeInt(data.readableBytes()+1);
 		raf.write(data);
 	}
